@@ -1,46 +1,22 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.11'
-
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+gem 'rails', '~> 4.1.4'
 
 gem 'pg'
 
-gem 'activeadmin'
+gem 'devise', '~> 3.2.4'
+gem 'activeadmin', github: 'gregbell/active_admin'
 
-
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-
-  # See https://github.com/sstephenson/execjs#readme for more supported runtimes
-  # gem 'therubyracer', :platforms => :ruby
-
-  gem 'uglifier', '>= 1.0.3'
-end
+gem 'sass-rails', '~> 4.0.3'
+gem 'coffee-rails', '~> 4.0.1'
+gem 'uglifier', '>= 2.5.3'
 
 group :development do
   gem 'pry'
+  gem 'annotate', git: 'git://github.com/ctran/annotate_models.git'
 end
 
-gem 'delayed_job_active_record'
+gem 'delayed_job_active_record', '~> 4.0.1'
 
-gem 'unicorn'
-
-# To use ActiveModel has_secure_password
-# gem 'bcrypt-ruby', '~> 3.0.0'
-
-# To use Jbuilder templates for JSON
-# gem 'jbuilder'
-
-# Use unicorn as the app server
-# gem 'unicorn'
-
-# Deploy with Capistrano
-# gem 'capistrano'
-
-# To use debugger
-# gem 'debugger'
+gem 'unicorn', '~> 4.8.3'
+gem 'rails_12factor', group: :production
