@@ -1,10 +1,12 @@
-ActiveAdmin.register AdminUser do     
+ActiveAdmin.register AdminUser do
+  permit_params :email, :password, :password_confirmation
+
   index do                            
     column :email                     
     column :current_sign_in_at        
     column :last_sign_in_at           
     column :sign_in_count             
-    default_actions                   
+    actions
   end                                 
 
   filter :email                       
